requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
dateparser==0.7.6
discord.py==1.4.1
idna==2.10
multidict==4.7.6
psutil==5.7.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.10.11
six==1.15.0
tzlocal==2.1
yarl==1.5.1
